[lab01] changes2

+++ util/helpers.h
@@ -4,7 +4,6 @@
 
 #ifndef LAB01_HELPERS_H
 #define LAB01_HELPERS_H
-#include <stdint.h>
 
 #define BUFSIZE 128
 #define DEFAULT_N_TIMES 1
@@ -14,8 +13,8 @@ int msleep(long msec);
 
 struct t_env_vars {
     long variant;
-    int msg_size;
-    uint16_t n;
+    long msg_size;
+    long n;
 };
 
 int read_env(struct t_env_vars *str);
